Update ca-on-kingston.json
Added website link. There's an Addresses section but the data seems to be provided on a per request basis with possible fees and restrictive licensing. I'll contact them to see if there's an easier way.

Quote from their site : Does the City have an Open Data policy?
No, not currently. However, the City is working on the development of an Open Data Policy for consideration by Council in the spring of 2014.