Update us-mn-ramsey.json
I am working on getting official permission on this and a better layer, but in the meantime I wanted to do the pull request so that folks knew I was tracking it down. And, I hope I can figure out how to do the conform object in order to do the coordinate and attribute translations.