Update ramsey.json
PostalCity, while not a "situs address city" but more like a "mailing address city", seems a better choice for this service, because the CityCode makes use of a coded value domain. In other words, a human-readable city doesn't come across, instead a numeric code. This change resolves that, although the "city" listed will often be for bulk-mailing purposes, not the actual city.