Update anoka.json with all caps field names
Looking at this REST endpoint and the resulting download here in OA, it looks to me like the REST query is case sensitive for the column names. The download appears successful for the column names where all caps was used, so I'm suggesting using all caps in all the column names.