Update saskatoon.json
The Saskatoon data is not being processed correctly because of html embedded in the kml file. I've parsed it with BeautifulSoup and have a temporary csv on the CSBP Github, hopefully this works until a more permanent processing solution can be developed.
This is also my 2nd attempt to merge this branch, forgot to specify the geocoordinates last time.