Update statewide.json (#6954)
* Update statewide.json

(re-done to update FILE attribute as requested) The path to the CO Statewide address point file has been updated. There are issues with the old file that have been corrected in this file. These issues go upstream to Geocode.Earth, for example, and are used by our District's School Finder.

* Specify layer

---------

Co-authored-by: archpdx <archpdx@pm.me>