Italy countrywide
This adds the processed data from https://github.com/ivandorte/anncsu_dump/
with a join to an administrative ISTAT table here https://www.istat.it/classificazione/codici-dei-comuni-delle-province-e-delle-regioni/ (also CC-BY 4.0) to add city, district, and region information.

This data still lacks zipcodes so is not a total upgrade from the previous CC BY SA version, but the freshness and massive increase in data seems worth the update.