Greenland addresses
Full country coverage

I didn't do a ton of research but there is both an address number and a building number as part of addressing. I added the building number as the unit for now so its accounted for.

The road, city, and region names were on a few different mapserver layers. that had to be joined through ids. I included a chatgpt script to assemble it all into one csv for future updates.