Taiwan regional updates
Updates most of the regional datasets for Taiwan. Kaohsiung, Penghu, Pingtung, Hsinchu county, and Hsinchu city still need updating but the sources are timing out currently.

I added a script to process and join the data to the city and region names. It catches and normalizes all the current quirks of the datasets as, despite being somewhat standardized, this data is inconsistent and messy. It wouldn't shock me though if the script needs further refining next refresh, but it should help speed up the process.